Rename rsciio.utils.utils_readfile to rsciio.utils.readfile
