+++ rsciio/digitalsurf/_api.py
@@ -42,7 +42,7 @@ import warnings
 
 # Maybe later we can implement reading the class with the io utils tools instead
 # of re-defining read functions in the class
-# import rsciio.utils.utils_readfile as iou
+# import rsciio.utils.readfile as iou
 
 # This module will prove useful when we write the export function
 # import rsciio.utils.tools
